Cambios del dia

+++ app/Http/Controllers/BillController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers;
 
 use \Carbon\Carbon;
 use App\Bill;
+use App\BillItem;
 use App\Company;
+use App\Provider;
 use App\Exports\BillExport;
+use App\Imports\BillImport;
 use Maatwebsite\Excel\Facades\Excel;
 use Illuminate\Http\Request;
 
@@ -30,7 +33,7 @@ class BillController extends Controller
     public function index()
     {
         $current_company = auth()->user()->companies->first()->id;
-        $bills = Bill::where('company_id', $current_company)->orderBy('generated_date', 'DESC')->paginate(10);
+        $bills = Bill::where('company_id', $current_company)->with('provider')->sortable(['generated_date' => 'desc'])->paginate(10);
         return view('Bill/index', [
           'bills' => $bills
         ]);
@@ -72,11 +75,46 @@ class BillController extends Controller
         $bill->other_reference = $request->other_reference;
         $bill->hacienda_status = "01";
         $bill->payment_status = "01";
-        $bill->payment_receipt = "VOUCHER-123451234512345";
+        $bill->payment_receipt = "";
         $bill->generation_method = "M";
       
         //Datos de proveedor
-        $bill->provider_id = $request->provider_id;
+        if( $request->provider_id == '-1' ){
+            $tipo_persona = $request->tipo_persona;
+            $identificacion_provider = $request->id_number;
+            $codigo_provider = $request->code;
+            
+            $provider = Provider::firstOrCreate(
+                [
+                    'id_number' => $identificacion_provider,
+                    'company_id' => $company->id,
+                ],
+                [
+                    'code' => $codigo_provider ,
+                    'company_id' => $company->id,
+                    'tipo_persona' => $tipo_persona,
+                    'id_number' => $identificacion_provider
+                ]
+            );
+            $provider->first_name = $request->first_name;
+            $provider->last_name = $request->last_name;
+            $provider->last_name2 = $request->last_name2;
+            $provider->country = $request->country;
+            $provider->state = $request->state;
+            $provider->city = $request->city;
+            $provider->district = $request->district;
+            $provider->neighborhood = $request->neighborhood;
+            $provider->zip = $request->zip;
+            $provider->address = $request->address;
+            $provider->phone = $request->phone;
+            $provider->es_exento = $request->es_exento;
+            $provider->email = $request->email;
+            $provider->save();
+                
+            $bill->provider_id = $provider->id;
+        }else{
+            $bill->provider_id = $request->provider_id;
+        }
         
         //Datos de factura
         $bill->description = $request->description;
@@ -139,8 +177,8 @@ class BillController extends Controller
         $bill = Bill::findOrFail($id);
         $this->authorize('update', $bill);
       
-        //Valida que la factura emitida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
-        if( $bill->generation_method != 'M' ){
+        //Valida que la factura recibida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
+        if( $bill->generation_method != 'M' && $bill->generation_method != 'XLSX' && $bill->generation_method != 'XML' ){
           return redirect('/facturas-recibidas');
         }  
       
@@ -160,10 +198,10 @@ class BillController extends Controller
         $bill = Bill::findOrFail($id);
         $this->authorize('update', $bill);
       
-        //Valida que la factura emitida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
-        if( $bill->generation_method != 'M' ){
+        //Valida que la factura recibida sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
+        if( $bill->generation_method != 'M' && $bill->generation_method != 'XLSX' && $bill->generation_method != 'XML' ){
           return redirect('/facturas-recibidas');
-        }
+        } 
       
         $company = $bill->company;
       
@@ -173,14 +211,44 @@ class BillController extends Controller
         $bill->credit_time = $request->credit_time;
         $bill->buy_order = $request->buy_order;
         $bill->other_reference = $request->other_reference;
-        $bill->hacienda_status = "01";
-        $bill->payment_status = "01";
-        $bill->payment_receipt = "VOUCHER-123451234512345";
-        $bill->generation_method = "M";
       
         //Datos de proveedor
-        $bill->provider_id = $request->provider_id;
-        $bill->send_emails = $request->send_emails;
+        if( $request->provider_id == '-1' ){
+            $tipo_persona = $request->tipo_persona;
+            $identificacion_provider = $request->id_number;
+            $codigo_provider = $request->code;
+            
+            $provider = Provider::firstOrCreate(
+                [
+                    'id_number' => $identificacion_provider,
+                    'company_id' => $company->id,
+                ],
+                [
+                    'code' => $codigo_provider ,
+                    'company_id' => $company->id,
+                    'tipo_persona' => $tipo_persona,
+                    'id_number' => $identificacion_provider
+                ]
+            );
+            $provider->first_name = $request->first_name;
+            $provider->last_name = $request->last_name;
+            $provider->last_name2 = $request->last_name2;
+            $provider->country = $request->country;
+            $provider->state = $request->state;
+            $provider->city = $request->city;
+            $provider->district = $request->district;
+            $provider->neighborhood = $request->neighborhood;
+            $provider->zip = $request->zip;
+            $provider->address = $request->address;
+            $provider->phone = $request->phone;
+            $provider->es_exento = $request->es_exento;
+            $provider->email = $request->email;
+            $provider->save();
+                
+            $bill->provider_id = $provider->id;
+        }else{
+            $bill->provider_id = $request->provider_id;
+        }
         
         //Datos de factura
         $bill->description = $request->description;
@@ -243,6 +311,11 @@ class BillController extends Controller
         $bill = Bill::find($id);
         $this->authorize('update', $bill);
         
+        //Valida que la factura sea generada manualmente. De ser generada por XML o con el sistema, no permite edición.
+        if( $bill->generation_method != 'M' && $bill->generation_method != 'XLSX' && $bill->generation_method != 'XML' ){
+          return redirect('/facturas-recibidas');
+        } 
+        
         foreach ( $bill->items as $linea ) {
           $linea->delete();
         }
@@ -250,13 +323,147 @@ class BillController extends Controller
         return redirect('/facturas-recibidas');
     }
     
-    public function export() {
-        return Excel::download(new InvoiceExport(), 'documentos-emitidos.xlsx');
-    }
     
+    public function export() {
+        return Excel::download(new BillExport(), 'documentos-emitidos.xlsx');
+    }
+
     public function import() {
         
-    }      
+        request()->validate([
+          'archivo' => 'required',
+          'tipo_archivo' => 'required',
+        ]);
+      
+        $time_start = $this->microtime_float();
+        
+        $facturas = Excel::toCollection( new BillImport(), request()->file('archivo') );
+        $company = auth()->user()->companies->first();
+        
+        foreach ($facturas[0] as $row){
+              
+            //Datos de proveedor
+            $codigo_proveedor = $row['codigoproveedor'] ? $row['codigoproveedor'] : '';
+            $nombre_proveedor = $row['nombreproveedor'];
+            $tipo_persona = $row['tipoidentificacion'];
+            $identificacion_proveedor = $row['identificacionproveedor'];
+            $proveedor = Provider::firstOrCreate(
+                [
+                    'id_number' => $identificacion_proveedor,
+                    'company_id' => $company->id,
+                ],
+                [
+                    'code' => $codigo_proveedor ,
+                    'company_id' => $company->id,
+                    'tipo_persona' => str_pad($tipo_persona, 2, '0', STR_PAD_LEFT),
+                    'id_number' => $identificacion_proveedor,
+                    'first_name' => $nombre_proveedor
+                ]
+            );
+                
+            $bill = Bill::firstOrNew(
+                [
+                    'company_id' => $company->id,
+                    'provider_id' => $proveedor->id,
+                    'document_number' => $row['consecutivocomprobante']
+                ]
+            );
+            
+            if( !$bill->exists ) {
+                
+                $bill->company_id = $company->id;
+                $bill->provider_id = $proveedor->id;    
+        
+                //Datos generales y para Hacienda
+                $bill->document_type = $row['idtipodocumento'];
+                $bill->reference_number = $company->last_bill_ref_number + 1;
+                $bill->document_number =  $row['consecutivocomprobante'];
+                
+                //Datos generales
+                $bill->sale_condition = str_pad($row['condicionventa'], 2, '0', STR_PAD_LEFT);
+                $bill->payment_type = str_pad($row['metodopago'], 2, '0', STR_PAD_LEFT);
+                $bill->credit_time = 0;
+                
+                /*$bill->buy_order = $row['ordencompra'] ? $row['ordencompra'] : '';
+                $bill->other_reference = $row['referencia'] ? $row['referencia'] : '';
+                $bill->other_document = $row['documentoanulado'] ? $row['documentoanulado'] : '';
+                $bill->hacienda_status = $row['estadohacienda'] ? $row['estadohacienda'] : '01';
+                $bill->payment_status = $row['estadopago'] ? $row['estadopago'] : '01';
+                $bill->payment_receipt = $row['comprobantepago'] ? $row['comprobantepago'] : '';*/
+                
+                $bill->generation_method = "XLSX";
+                
+                //Datos de factura
+                $bill->currency = $row['idmoneda'];
+                if( $bill->currency == 1 ) { $bill->currency = "CRC"; }
+                if( $bill->currency == 2 ) { $bill->currency = "USD"; }
+                    
+                
+                $bill->currency_rate = $row['tipocambio'];
+                //$bill->description = $row['description'] ? $row['description'] : '';
+              
+                $company->last_bill_ref_number = $bill->reference_number;
+                
+                $bill->subtotal = 0;
+                $bill->iva_amount = 0;
+                $bill->total = $row['totaldocumento'];
+                
+                $bill->save();
+            }     
+            
+          
+            /**LINEA DE FACTURA**/
+            $item = BillItem::firstOrNew(
+                [
+                    'bill_id' => $bill->id,
+                    'item_number' => $row['numerolinea'],
+                ],
+                [
+                'bill_id' => $bill->id,
+                'item_number' => $row['numerolinea'],
+                'code' => $row['codigoproducto'],
+                'name' => $row['detalleproducto'],
+                'product_type' => 1,
+                'measure_unit' => $row['unidadmedicion'],
+                'item_count' => $row['cantidad'],
+                'unit_price' => $row['preciounitario'],
+                'subtotal' => $row['subtotallinea'],
+                'total' => $row['totallinea'],
+                'discount_type' => '01',
+                'discount' => $row['montodescuento'],
+                'discount_reason' => '',
+                'iva_type' => $row['codigoimpuesto'],
+                'iva_amount' => $row['montoiva'],
+                ]
+            );
+            
+            if( !$item->exists ) {
+                $bill->subtotal = $bill->subtotal + (float)$row['subtotallinea'];
+                $bill->iva_amount = $bill->iva_amount + (float)$row['montoiva'];
+                $item->save();
+            }
+            /**END LINEA DE FACTURA**/
+            
+            $bill->generated_date = Carbon::createFromFormat('d/m/Y', $row['fechaemision']);
+            $bill->due_date = Carbon::createFromFormat('d/m/Y', $row['fechaemision'])->addDays(15);
+            
+            
+            
+            $bill->save();
+            
+        }
+        $company->save();
+        
+        $time_end = $this->microtime_float();
+        $time = $time_end - $time_start;
+        
+        return redirect('/facturas-recibidas')->withMessage('Facturas importados exitosamente en '.$time.'s');
+    }
+    
+    private function microtime_float(){
+        list($usec, $sec) = explode(" ", microtime());
+        return ((float) $usec + (float)$sec);
+    }   
     
     
 }
